Add files via upload
There are so many confusions happening due to incorrect upload of signature and photo. So this model ensures that they are uploaded correctly by classifying them into signature and photo. This repository contains the codes that train the model by tensorflow, scikit-learn and custom basic neural network and a trained model in sav file.